+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1. 兼容 eslint v7.x 版本
 2. `@eslint-sets/eslint-config-all` 更改为 `@eslint-sets/eslint-config`
+3. `@nuxtjs/eslint-config` 升级到 v12
 
 ## 2023.01.03 v3.3.0 更新日志